requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
pillow==10.0.0
python-chess==1.10.0
scikit-image==0.21.0
matplotlib==3.7.2
scipy==1.11.3
